digital-time-capsule/requirements.txt
flask
flask-cors
pymongo
python-dotenv
flask-pymongo
gunicorn
google-cloud-storage
google-cloud-firestore